requirements.txt
openai
discord.py
python-dotenv
pillow